requirements.txt
pandas>=2.2.0
pika>=1.3.2
python-dotenv>=1.0.1
psycopg2-binary>=2.9.0
pytest>=8.2.0

